% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{sunsetrise_yr}
\alias{sunsetrise_yr}
\title{sunsetrise_yr}
\usage{
sunsetrise_yr(year, lon_sign = -1, lon_deg = NA, lon_min = NA,
  lat_sign = 1, lat_deg = NA, lat_min = NA, tz = NA, tz_sign = 1,
  state = NA, city = NA)
}
\arguments{
\item{year}{integer}

\item{lon_sign}{integer -1 for west}

\item{lon_deg}{integer degree}

\item{lon_min}{integer minute}

\item{lat_sign}{integer -1 for south}

\item{lat_deg}{integer degree}

\item{lat_min}{integer minute}

\item{tz}{integer timezone}

\item{tz_sign}{integer}

\item{state}{character two letter United States state abbreviation}

\item{city}{character}
}
\value{
A 3 x 365 data.frame of sunrise, sunset, date where times are in GMT
}
\description{
sunsetrise_yr
}
\examples{
\dontrun{
sunsetrise_yr(year = 2015, lon_deg = 80, lon_min = 26, lat_deg = 25, lat_min = 5, tz = 5)

sunsetrise_yr(year = 2015, state = "FL", city = "Key Largo")
}
}
\references{
\url{http://aa.usno.navy.mil/data/docs/RS_OneYear.php}
}

