% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{tw_daylength}
\alias{tw_daylength}
\title{tw_daylength}
\usage{
tw_daylength(...)
}
\arguments{
\item{...}{arguments passed to sunsetrise_yr}
}
\description{
Daylength in hhmm format
}
\examples{
\dontrun{
tw_daylength(year = 2015, lon_deg = 80, lon_min = 26, lat_deg = 25, lat_min = 5, tz = 5)

tw_daylength(year = 2015, state = "FL", city = "Key Largo")
}
}

