% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilight-package.R
\docType{package}
\name{twilight-package}
\alias{twilight}
\alias{twilight-package}
\title{R interface to the USNO Sun or Moon Rise/Set Data Service}
\description{
R interface to the USNO Sun or Moon Rise/Set Data Service
}
\author{
\email{stachel2@msu.edu}
}

