\name{exfdr}
\alias{exfdr}
\docType{data}
\title{ Example of twilight result. }
\description{
Application of function \code{twilight} on twilight object \code{data(expval)}.

The function call was \code{exfdr <- twilight(expval,B=1000)}.
}
\usage{data(exfdr)}
\format{A twilight object.}
\keyword{datasets}
